With BS410 Send correct time to plugins
When using BS410 not only send the correct time to the device but also
correct (i.e. convert to unix timestamp) the received timestamp as sent
to the plugins. Also got rid of decode.py as a separate file. Serves no
purpose to separate it from the rest of the code.
